Implemented switchanble event/form
