+++ src/components/trip-event-edit.js
@@ -34,7 +34,8 @@ export const tripEventEdit = () => {
     getTemplate() {
       const price = this._optionCost[0] + this._intermadiatePrice;
 
-      return `<form class="event  event--edit" action="#" method="post">
+      return `<li class="trip-events__item">
+      <form class="event  event--edit" action="#" method="post">
         <header class="event__header">
           <div class="event__type-wrapper">
             <label class="event__type  event__type-btn" for="event-type-toggle-1">
@@ -193,7 +194,8 @@ export const tripEventEdit = () => {
             </div>
           </section>
         </section>
-      </form>`;
+      </form>
+    </li>`;
     }
   }
